Fixes and pre-pull of 2.3.4
Fixed delay when loading a savestate if music is disabled. Still delays if it has to load new music
Fixed background changing colors if the practice menu is opened while the game is paused
Getting closer to release, this could be it. The rest of this is from idlechild.
Load presets in Ceres Ridley's room
Updates to music loading during load state and preset load