Precise Phantoon RNG by NobodyNada
Locked behind a build option for now: !NEW_PHANTOON_RNG

Add hooks to Phantoon's AI that override the default RNG-based
pattern chooser in order to limit patterns to user-selected subset.

If the user enabled all the patterns (or none of them), Phantoon's
regular RNG code is run instead -- so this feature should not impact
Phantoon's behavior or pattern probabilities when not in use.

Co-Authored-By: Jonathan Keller <19418817+NobodyNada@users.noreply.github.com>