Fix presets that load into shinesparks
The spark timer must be set to at least 2 or it will be decremented to zero or negative before the spark can get started. The timer should be set high in case the player holds inputs after loading, which would allow the timer to decrement before normal shinespark code is running. This fixes unintentional spikesuits from presets saved mid-spark.